+++ src/main/java/fuzs/deathfinder/DeathFinder.java
@@ -3,12 +3,16 @@ package fuzs.deathfinder;
 import fuzs.deathfinder.config.ClientConfig;
 import fuzs.deathfinder.config.ServerConfig;
 import fuzs.deathfinder.handler.DeathMessageHandler;
+import fuzs.deathfinder.network.chat.AdvancedClickEvent;
+import fuzs.deathfinder.network.chat.TeleportClickEvent;
 import fuzs.deathfinder.network.client.message.C2SDeathPointTeleportMessage;
+import fuzs.deathfinder.network.message.S2CAdvancedChatMessage;
 import fuzs.deathfinder.registry.ModRegistry;
 import fuzs.puzzleslib.config.ConfigHolder;
 import fuzs.puzzleslib.config.ConfigHolderImpl;
 import fuzs.puzzleslib.network.MessageDirection;
 import fuzs.puzzleslib.network.NetworkHandler;
+import net.minecraft.resources.ResourceLocation;
 import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
@@ -33,6 +37,7 @@ public class DeathFinder {
         registerHandlers();
         registerMessages();
         ModRegistry.touch();
+        AdvancedClickEvent.register(new ResourceLocation(MOD_ID, "teleport"), TeleportClickEvent.class, TeleportClickEvent::new);
     }
 
     private static void registerHandlers() {
@@ -41,6 +46,7 @@ public class DeathFinder {
     }
 
     private static void registerMessages() {
+        NETWORK.register(S2CAdvancedChatMessage.class, S2CAdvancedChatMessage::new, MessageDirection.TO_CLIENT);
         NETWORK.register(C2SDeathPointTeleportMessage.class, C2SDeathPointTeleportMessage::new, MessageDirection.TO_SERVER);
     }
 }