+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v3.1.1-1.18.1] - 2022-02-24
+### Changed
+- Teleporting to death positions is no longer enabled for normal players by default
+### Fixed
+- Teleporting to death points now properly works for all players on dedicated servers (when enabled)
+
 ## [v3.1.0-1.18.1] - 2022-02-22
 ### Changed
 - Moved internal capability system to Puzzles Lib